+++ frontend/src/hooks/useApi.ts
@@ -52,11 +52,6 @@ export default function useApi() {
   }
 
   /**
-   * User Api
-   * ##########################################################################
-   */
-
-  /**
    * Season Api
    * ##########################################################################
    */
